# services/mcp_server/app/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Enhanced MCP API",
    "description": "Example tools for MCP integration.",
    "version": "1.0.0"
  },
  "servers": [{"url": "http://localhost:8090"}],
  "paths": {
    "/tools/get_vehicle_price": {
      "post": {
        "operationId": "get_vehicle_price",
        "summary": "Get vehicle price",
        "description": "Returns a price for the given brand and model.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "brand": {"type": "string", "description": "Vehicle brand"},
                  "model": {"type": "string", "description": "Vehicle model"}
                },
                "required": ["brand", "model"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "brand": {"type": "string"},
                    "model": {"type": "string"},
                    "price_eur": {"type": "integer"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tools/get_weather": {
      "post": {
        "operationId": "get_weather",
        "summary": "Get weather",
        "description": "Return current weather for a location.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "location": {"type": "string", "description": "City or region"},
                  "units": {"type": "string", "description": "metric or imperial", "default": "metric"}
                },
                "required": ["location"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "location": {"type": "string"},
                    "temperature": {"type": "number"},
                    "units": {"type": "string"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tools/calculate": {
      "post": {
        "operationId": "calculate",
        "summary": "Calculate expression",
        "description": "Evaluate a simple arithmetic expression.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "expression": {"type": "string", "description": "Math expression"}
                },
                "required": ["expression"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {"type": "number"}
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}